Split request ID tagging into a separate, earlier filter
To allow correlating database queries with request IDs, move the code
that populates the request context with a random request ID to a
separate filter that runs earlier in the filter chain. This way the
request ID can be included with any database queries done by other
filters (such as authentication).
